+++ src/main/java/com/ktb/cafeboo/global/infra/ai/dto/CreateWeeklyAnalysisResponse.java
@@ -1,0 +1,24 @@
+package com.ktb.cafeboo.global.infra.ai.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.Data;
+import lombok.Getter;
+
+@Data
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class CreateWeeklyAnalysisResponse {
+    private String status;
+    private String message;
+    private Data data;
+
+    @Getter
+    public static class Data {
+        @JsonProperty("user_id")
+        private String userId;
+        private String report;
+    }
+}
